pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
    packages = ["src/**/*"]

[tool.hatch.metadata]
allow-direct-references = true

[project]
name = "latent_space_explorer"
version = "0.1.0"
description = "Latent Space Explorer"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.11"

dependencies = [
    "dash==2.9.3",
    "dash-core-components==2.0.0",
    "dash-bootstrap-components>=1.0.2",
    "dash-extensions==0.0.71",
    "dash-iconify==0.1.2",
    "flask==3.0.0",
    "Flask-Caching",
    "kaleido",
    "humanhash3",
    "mlex_file_manager@git+https://github.com/mlexchange/mlex_file_manager.git",
    "mlex_utils[all]@git+https://github.com/mlexchange/mlex_utils.git",
    "numpy>=1.19.5",
    "pandas",
    "Pillow",
    "plotly>=5.21.0",
    "plotly-express",
    "pyFAI==2023.9.0",
    "python-dotenv",
    "pyarrow==11.0.0",
    "requests==2.26.0",
    "diskcache==5.6.3",
    "scikit-learn==1.3.0",
]

[project.optional-dependencies]
dev = [
    "black==24.2.0",
    "flake8==7.0.0",
    "isort==5.13.2",
    "pre-commit==3.6.2",
    "pytest==8.1.1",
]

simulator = [
    "websocket",
    "pika",
]
